spark-pipeline/requirements.txt
pyspark==3.5.0
pg8000==1.29.8
pandas==2.1.1
confluent-kafka==2.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
